+++ src/pyjams/__init__.py
@@ -31,9 +31,9 @@ Subpackages
    morris_method
    position
    screening
-   signature2plot
    str2tex
    tee
+   text2plot
 
 
 History
@@ -49,7 +49,7 @@ History
     * v1.8, added str2tex and color, Nov 2021, Matthias Cuntz
     * v1.9, added position, Nov 2021, Matthias Cuntz
     * v1.10, added sron colors, Nov 2021, Matthias Cuntz
-    * v1.11, added signature2plot, Nov 2021, Matthias Cuntz
+    * v1.11, added abc2plot, signature2plot, Nov 2021, Matthias Cuntz
 
 """
 # version, author
@@ -89,8 +89,8 @@ from .position import position
 from .screening import screening, ee
 # Convert strings to LaTeX strings
 from .str2tex import str2tex
-# put copyright notice on plot (import after str2tex)
-from .signature2plot import signature2plot
+# put text on plot (import after str2tex)
+from .text2plot import abc2plot, signature2plot
 
 # like unix tee
 from .tee import tee
@@ -105,6 +105,6 @@ __all__ = ['__version__', '__author__',
            'mcPlot',
            'morris_sampling', 'elementary_effects',
            'screening', 'ee',
-           'signature2plot'
+           'abc2plot', 'signature2plot',
            'tee',
            ]